Make Sonar happy

+++ app/src/main/java/me/devsaki/hentoid/fragments/queue/QueueFragment.java
@@ -440,7 +440,7 @@ public class QueueFragment extends Fragment implements ItemTouchCallback, Simple
             dlPreparationProgressBar.setVisibility(View.GONE);
         }
 
-        dlPreparationProgressBar.setProgress1(event.total - event.done);
+        dlPreparationProgressBar.setProgress1((long)event.total - event.done);
     }
 
     /**
